More restrucuring.
