+++ tests/test_budget_equation.py
