+++ tests/utils/upper_envelope_fedor.py
@@ -11,7 +11,6 @@ from typing import Tuple
 
 import numpy as np
 from dcegm.interpolation import linear_interpolation_with_extrapolation
-from dcegm.interpolation import linear_interpolation_with_inserting_missing_values
 from scipy.optimize import brenth as root
 
 eps = 2.2204e-16
@@ -648,3 +647,28 @@ def _subtract_values(grid_point: float, first_segment, second_segment):
     diff_values_segments = values_first_segment - values_second_segment
 
     return diff_values_segments
+
+
+def linear_interpolation_with_inserting_missing_values(x, y, x_new, missing_value):
+    """Linear interpolation with inserting missing values.
+
+    Args:
+        x (np.ndarray): 1d array of shape (n,) containing the x-values.
+        y (np.ndarray): 1d array of shape (n,) containing the y-values
+            corresponding to the x-values.
+        x_new (np.ndarray or float): 1d array of shape (m,) or float containing
+            the new x-values at which to evaluate the interpolation function.
+        missing_value (np.ndarray or float): Flat array of shape (1,) or float
+            to set for values of x_new outside of the range of x.
+
+    Returns:
+        np.ndarray or float: 1d array of shape (m,) or float containing the
+            new y-values corresponding to the new x-values.
+            In case x_new contains values outside of the range of x, these
+            values are set equal to missing_value.
+
+    """
+    interpol_res = linear_interpolation_with_extrapolation(x, y, x_new)
+    where_to_miss = (x_new < x.min()) | (x_new > x.max())
+    interpol_res[where_to_miss] = missing_value
+    return interpol_res
